requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
PyMySQL==1.1.1
python-decouple==3.8
PyYAML==6.0.1
requests==2.32.3
setuptools==69.5.1
SQLAlchemy==2.0.30
typing_extensions==4.12.1
urllib3==2.2.1
wheel==0.43.0
